+++ mcww_web/js/gallery.js
@@ -275,3 +275,15 @@ onUiUpdate((updatedElements) => {
         });
     });
 });
+
+
+function getGalleryContainerUnderCursor() {
+    const lastMouseEvent = getLastMouseEvent();
+    const elementUnderCursor = document.elementFromPoint(lastMouseEvent.clientX, lastMouseEvent.clientY);
+    let container = null;
+    if (elementUnderCursor) {
+        container = elementUnderCursor.closest('.gallery-container, .image-container, .video-container, .mcww-other-gallery, .upload-gallery');
+    }
+    return container;
+}
+