+++ Changelog.md
@@ -5,6 +5,7 @@
 - Fix for nodes with fields after dynamic field (fixes the updated text generation node with use_default_template and thinking); Fixed input video saving; Fixed subgraphs with proxy widgets
 - Support for `<think></think>` inside markdown view
 - Can open file on drop anywhere in the document
+- Can open file (image) on Ctrl+V in the document
 - Can open 3D models files, so you can use MCWW as a simple 3D models viewer. It doesn't support nodes with 3D models in workflows yet. If you installed the PWA before, you need to reinstall it (with cleaned browse's cache) to register extensions associations with these files
 - Can open markdown files - use MCWW as a simple markdown viewer
 - Split presets dataset in pages for performance reasons. If there are too many presets, workflow rendering takes too much time. Default page size is 50, can be changed in the settings